% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomizeStarz.R
\name{dichotomizeStarz}
\alias{dichotomizeStarz}
\title{Dichotomize Starz}
\usage{
dichotomizeStarz(star_vector)
}
\arguments{
\item{star_vector}{a vector of integers valued between 1 and 5}
}
\value{
a vector of 0's and 1's; the dichomomized rating
}
\description{
Turns the multinomial ratings into a binomial rating
}
\details{
Rating values can ONLY be 1,2,3,4,5. Values of 1,2,3 are 
given a new rating of 0, values or 4,5 are given a new 
rating of 1
}
