% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCountData.R
\name{getCountData}
\alias{getCountData}
\title{Get Count Data}
\usage{
getCountData(sentence_vector)
}
\arguments{
\item{sentence_vector}{a vector of sentences - factor or character - that
we wish to analyze.}
}
\value{
a data frame with columns: wd_ct, pos_ct, pos_prop, neg_ct, neg_prop. Which
are the number of words, number of positive words, proprtion of positive words, 
number of negative words and proportion of negative words in the sentence, respectively.
}
\description{
Gets the number of words, number/proportion of positive words,
and number/proportion of negative words for each sentence in the
sentence vector.
}
\details{
First, we clean each sentence in the vector, then we get the word count,
and then we get the number/proportion of positive words,
and number/proportion of negative words for each sentence in the
sentence vector.

This function is an envelope function for cleanSentence(), wordcount(), getPos(),
getNeg(), and getSentCount(). This function just makes it easier to use everything
in one fell swoop for the purposes of our project. You can use each of the functions
independently if you would rather - especially if you would like to supply your own
vector of positive and negative words.
}
