% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNeg.R
\name{getNeg}
\alias{getNeg}
\title{Get Negative Words}
\usage{
getNeg()
}
\value{
returns a vector of the negative words
}
\description{
Gets the negative words based on the bing definition in
get_sentiments()
}
\details{
Creates a data frame of the bing words and their corresponding 
sentiment which can be either positive or negative. Then it
creates a vector of only the negative words.
}
