% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPos.R
\name{getPos}
\alias{getPos}
\title{Get Positive Words}
\usage{
getPos()
}
\value{
returns a vector of the positive words
}
\description{
Gets the positive words based on the bing definition in
get_sentiments()
}
\details{
Creates a data frame of the bing words and their corresponding 
sentiment which can be either positive or negative. Then it
creates a vector of only the positive words.
}
