% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSentCount.R
\name{getSentCount}
\alias{getSentCount}
\title{Get Sentiment Count}
\usage{
getSentCount(clean_sentence, wdct = ngram::wordcount(clean_sentence),
  pos_words = getPos(), neg_words = getNeg())
}
\arguments{
\item{clean_sentence}{a single sentence that has been standardized}

\item{wdct}{the number of words in the sentence. Default is to calculate
word count if it wasn't supplied}

\item{pos_words}{a vector of words that are considered positive. Default
is to use the bing definition in the get_sentiments() function.}

\item{neg_words}{a vector of words that are considered negative. Default
is to use the bing definition in the get_sentiments() function.}
}
\value{
a vector with number of positive words, proportion of positive 
words,number of negative words, and the proportion of negative words.
}
\description{
For a given, clean, sentence, it gets the number of positive 
 and number of negative words in the sentence.
}
\details{
First, if the word count is greater than zero, we create a
vector of words in the sentence. Then we count the number of
words in the sentence that are in the positive vector and the 
number that are in the negative vector. We also calculate the
proportion of positive words and proportion of negative words.

The default vector of positive and negative words are from the
bing definition in the get_sentiments() function. However, 
users are welcome to supply their own vector of words if they
prefer.

If the word count is zero, it returns all missing values
}
