% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQA.lasso.R
\name{LQA.lasso}
\alias{LQA.lasso}
\title{LQA Lasso}
\usage{
LQA.lasso(X, Y, beta_start, lambda, tol, max_it)
}
\arguments{
\item{X}{design matrix with an intercept column and numeric variables that have been centered and scaled}

\item{Y}{vector of binary outcomes taking values 0 and 1}

\item{beta_start}{vector of starting values for the parameters, should have length equal to number of columns in X}

\item{lambda}{scalar value of the penalty coefficient}

\item{tol}{desired tolerance level for convergence criteria}

\item{max_it}{maximum number of iterations}
}
\value{
a vector of parameter estimates beta_tilde, with length equal to the number of columns in X
}
\description{
This function performs LASSO regression given a design matrix, an outcome vector,
starting values for the regression parameters, and a penalty coefficient.
}
