% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.thresh.R
\name{soft.thresh}
\alias{soft.thresh}
\title{Soft Threshold}
\usage{
soft.thresh(z, lambda)
}
\arguments{
\item{z}{description}

\item{lambda}{description}
}
\value{
a scalar value
}
\description{
This function does...
}
\examples{
give examples

}
