% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LQA.lasso.R
\name{LQA.lasso}
\alias{LQA.lasso}
\title{LQA Lasso}
\usage{
LQA.lasso(X, Y, beta_start, lambda, tol, max_it)
}
\arguments{
\item{X}{description}

\item{Y}{description}

\item{beta_start}{description}

\item{lambda}{description}

\item{tol}{description}

\item{max_it}{description}
}
\value{
a vector of parameters beta_tilde
}
\description{
This function does...
}
\examples{
give examples

}
