% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.thresh.R
\name{soft.thresh}
\alias{soft.thresh}
\title{Soft Threshold}
\usage{
soft.thresh(z, lambda)
}
\arguments{
\item{z}{numeric value}

\item{lambda}{numeric value of the penalty coefficient}
}
\value{
a scalar value
}
\description{
This function produces a soft threshold value to be used in the LASSO function 
to compute the updated vector of parameter estimates.
}
